Update service and tests
