+++ app/services/bill-runs/start-bill-run-process.service.js
@@ -10,7 +10,7 @@ const DetermineBillingPeriodsService = require('./determine-billing-periods.serv
 const CreateBillRunPresenter = require('../../presenters/bill-runs/create-bill-run.presenter.js')
 const InitiateBillRunService = require('./initiate-bill-run.service.js')
 const SupplementaryProcessBillRunService = require('./supplementary/process-bill-run.service.js')
-const TwoPartTariffProcessBillRunService = require('./two-part-tariff/process-bill-run.service.js')
+const ProcessTwoPartTariffReturnsService = require('./two-part-tariff/process-two-part-tariff-returns.service.js')
 
 /**
  * Manages the creation of a new bill run
@@ -50,7 +50,7 @@ function _processBillRun (billRun, billingPeriods) {
       SupplementaryProcessBillRunService.go(billRun, billingPeriods)
       break
     case 'two_part_tariff':
-      TwoPartTariffProcessBillRunService.go(billRun, billingPeriods)
+      ProcessTwoPartTariffReturnsService.go(billRun, billingPeriods)
       break
   }
 }
