--- test/services/bill-runs/two-part-tariff/process-bill-run.service.test.js
@@ -1,24 +1,0 @@
-'use strict'
-
-// Test framework dependencies
-const Lab = require('@hapi/lab')
-const Code = require('@hapi/code')
-
-const { describe, it } = exports.lab = Lab.script()
-const { expect } = Code
-
-// Thing under test
-const TwoPartTariffProcessBillRunService = require('../../../../app/services/bill-runs/two-part-tariff/process-bill-run.service.js')
-
-describe('Two Part Tariff Process Bill Run service', () => {
-  describe('when the service is called', () => {
-    const billingPeriods = [{ endDate: new Date('2023-03-31') }]
-
-    it('throws an error', async () => {
-      const error = await expect(TwoPartTariffProcessBillRunService.go('billRun', billingPeriods)).to.reject()
-
-      expect(error).to.be.an.instanceOf(Error)
-      expect(error.message).to.equal('Two Part Tariff is not yet implemented for Financial Year Ending: 2023')
-    })
-  })
-})
